Rename "surrogates" directory as "functional"
