+++ src/sqlfluff/core/rules/functional/raw_file_slice.py
