+++ src/sqlfluff/core/rules/functional/segment_predicates.py
