+++ src/sqlfluff/core/rules/functional/segments.py
@@ -3,7 +3,7 @@ from typing import Any, Callable, List, Optional, Sequence, Type, Union
 
 from sqlfluff.core.parser import BaseSegment
 from sqlfluff.core.templaters.base import TemplatedFile
-from sqlfluff.core.rules.surrogates.raw_file_slice import RawFileSlices
+from sqlfluff.core.rules.functional.raw_file_slice import RawFileSlices
 
 Predicate = Union[str, Type, Callable[[BaseSegment], bool]]
