+++ src/sqlfluff/rules/L013.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from sqlfluff.core.rules.base import BaseRule, LintResult, RuleContext
 from sqlfluff.core.rules.doc_decorators import document_configuration
-import sqlfluff.core.rules.surrogates.segment_predicates as segpred
+import sqlfluff.core.rules.functional.segment_predicates as segpred
 
 
 @document_configuration
